extract Lit components from style into @commons-systems/components
Move hero, hero-render, panel-toggle, scroll-indicator, nav, and
autocomplete out of the style package into a new components package.
The style package becomes CSS-only (design tokens + generated CSS);
no runtime Lit code remains there.

Update all consumers (audio, budget, fellspiral, landing, print,
scaffolding template) to import from @commons-systems/components.
Add the new workspace package to the root package.json and
package-lock.json.
