+++ budget/src/main.ts
@@ -7,8 +7,8 @@ import { renderBudgets } from "./pages/budgets.js";
 import { renderAccounts } from "./pages/accounts.js";
 import { renderAccountsReconcile } from "./pages/accounts-reconcile.js";
 import { renderRules } from "./pages/rules.js";
-import "@commons-systems/style/components/nav";
-import type { AppNavElement } from "@commons-systems/style/components/nav";
+import "@commons-systems/components/nav";
+import type { AppNavElement } from "@commons-systems/components/nav";
 import type { RenderPageOptions } from "./pages/render-options.js";
 import { hydrateTransactionTable } from "./pages/home-hydrate.js";
 import { hydrateCategorySankey } from "./pages/home-chart.js";
@@ -16,7 +16,7 @@ import { hydrateBudgetTable, hydrateBudgetChart, hydrateOverridesTable } from ".
 import { hydrateRulesTable } from "./pages/rules-hydrate.js";
 import { hydrateAccountsCharts } from "./pages/accounts-hydrate.js";
 import { hydrateAccountsReconcile } from "./pages/accounts-reconcile-hydrate.js";
-import { mountHero } from "@commons-systems/style/hero";
+import { mountHero } from "@commons-systems/components/hero";
 import { renderHero } from "./pages/hero.js";
 import { trackPageView, initAppCheck } from "./firebase.js";
 import { classifyError } from "@commons-systems/errorutil/classify";
