+++ budget/src/pages/home-chart.ts
@@ -1,7 +1,7 @@
 import { hierarchy, tree, type HierarchyNode } from "d3-hierarchy";
 import { computeNetAmount, isCardPaymentCategory, MS_PER_WEEK, weekStart } from "../balance.js";
 import { formatCurrency } from "../format.js";
-import { showDropdown, registerAutocompleteListeners } from "@commons-systems/style/components/autocomplete";
+import { showDropdown, registerAutocompleteListeners } from "@commons-systems/components/autocomplete";
 import { parseJsonArray, makeDebounced } from "./hydrate-util.js";
 import { getThemeFg } from "./chart-util.js";
