+++ components/src/hero-render.ts
