+++ components/src/hero.ts
