+++ components/src/scroll-indicator.ts
