+++ components/test/scroll-indicator.test.ts
