+++ components/tsconfig.json
@@ -1,4 +1,7 @@
 {
   "extends": "@commons-systems/config/tsconfig.lib.json",
+  "compilerOptions": {
+    "lib": ["ES2022", "DOM", "DOM.Iterable"]
+  },
   "include": ["src", "test"]
 }
