+++ landing/src/main.ts
@@ -16,9 +16,9 @@ import { createFetchPost } from "@commons-systems/blog/github";
 import { updateOgMeta } from "@commons-systems/blog/og-meta";
 import { updateCanonical } from "@commons-systems/blog/canonical";
 import { getPosts, type PostMeta } from "@commons-systems/blog/firestore";
-import { initPanelToggle } from "@commons-systems/style/panel-toggle";
-import "@commons-systems/style/components/nav";
-import type { AppNavElement } from "@commons-systems/style/components/nav";
+import { initPanelToggle } from "@commons-systems/components/panel-toggle";
+import "@commons-systems/components/nav";
+import type { AppNavElement } from "@commons-systems/components/nav";
 import { BLOG_ROLL_ENTRIES, createStrategies } from "./blog-roll/config.js";
 import { INFO_PANEL_LINK_SECTIONS, SITE_DEFAULTS, SITE_URL } from "./site-config.js";
 import { signIn, signOut, onAuthStateChanged } from "./auth.js";
