+++ print/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@commons-systems/idbutil": "*",
     "@commons-systems/analyticsutil": "*",
+    "@commons-systems/components": "*",
     "@commons-systems/errorutil": "*",
     "@commons-systems/authutil": "*",
     "@commons-systems/firebaseutil": "*",
