+++ print/src/main.ts
@@ -5,13 +5,13 @@ import { classifyError } from "@commons-systems/errorutil/classify";
 import { renderHome, afterRenderHome } from "./pages/home.js";
 import { renderView, afterRenderView, cleanupView } from "./pages/view.js";
 import { renderAbout } from "./pages/about.js";
-import "@commons-systems/style/components/nav";
-import type { AppNavElement } from "@commons-systems/style/components/nav";
+import "@commons-systems/components/nav";
+import type { AppNavElement } from "@commons-systems/components/nav";
 import { signIn, signOut, onAuthStateChanged } from "./auth.js";
 import type { User } from "./auth.js";
 import { trackPageView } from "./firebase.js";
 import { renderHero } from "./pages/hero.js";
-import { mountHero } from "@commons-systems/style/hero";
+import { mountHero } from "@commons-systems/components/hero";
 
 const navEl = document.getElementById("nav") as AppNavElement;
 if (!navEl) throw new Error("#nav element not found");
