+++ scaffolding/firebase/templates/app/package.json.tmpl
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@commons-systems/analyticsutil": "file:../analyticsutil",
+    "@commons-systems/components": "file:../components",
     "@commons-systems/authutil": "file:../authutil",
     "@commons-systems/router": "file:../router",
     "@commons-systems/firebaseutil": "file:../firebaseutil",
